refactor: update file name of AceConfigurationPanel
